% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_group_summarise.R
\name{quick_group_summarise}
\alias{quick_group_summarise}
\title{Quick Group By and Summarise a Mean}
\usage{
quick_group_summarise(data, x, y)
}
\arguments{
\item{data}{The data set you would like to work with}

\item{x}{Any variable from your data set that you would like to examine}

\item{y}{Any numeric variable from your data set}
}
\value{
This function will return a tibble that shows the grouped by variable (x) and the corresponding mean of the numeric variable (y). There will only be two columns in this new tibble.
}
\description{
This function first groups by one variable in your data set (x). Then, it summarises the mean of a chosen numeric variable (y) for each value in the grouped by variable (x) and produces a new tibble.
}
\examples{
quick_group_summarise(datateachr::steam_games, genre, original_price)
#Should produce a tibble with the mean price for each genre

quick_group_summarise(datateachr::steam_games, languages, original_price)
#Should produce a tibble with mean price for each language
}
